requirements.txt
langchain_community
openai
python-dotenv
langchain-openai
faiss-cpu
langchain_core
streamlit
langchain_groq
fastapi
langserve
sse_starlette